React app that fetches data using axios and giphy API with a search function
React app that fetches data using axios and giphy API with a search
function that displays gifs with an entered (set of) keyword(s).

Application manages state across components, interacts with an API to
make a fetch request (the native fetch API is replaced in this
application with the axios library), and loading state is managed for
when the app starts and there is a brief time during which no gifs are
have been loaded.